more optimisations
