+++ apps/ui/registry/default/atoms/booker-1.tsx
@@ -1,14 +1,7 @@
 "use client";
 
 import { ArrowLeftIcon, Clock3Icon } from "lucide-react";
-import {
-  animate,
-  domAnimation,
-  LazyMotion,
-  MotionConfig,
-  useReducedMotion,
-} from "motion/react";
-import { type ReactNode, useEffect, useLayoutEffect, useRef } from "react";
+import { cn } from "@/registry/default/lib/utils";
 import { Button } from "@/registry/default/ui/button";
 import { Card } from "@/registry/default/ui/card";
 import { Input } from "@/registry/default/ui/input";
@@ -17,6 +10,7 @@ import { BookerAvatars } from "./booker/booker-avatars";
 import { BookerCalendar } from "./booker/booker-calendar";
 import { BookerErrorState } from "./booker/booker-error-state";
 import { type BookerLabels, getBookerLabels } from "./booker/booker-labels";
+import { BookerSteps } from "./booker/booker-steps";
 import { DurationPicker } from "./booker/duration-picker";
 import { EventDescription } from "./booker/event-description";
 import { HeaderBanner } from "./booker/header-banner";
@@ -24,118 +18,7 @@ import { Location } from "./booker/location";
 import { TimePicker } from "./booker/time-picker";
 import { TimezonePicker } from "./booker/timezone-picker";
 import type { BookerTarget } from "@/lib/booker/target";
-import {
-  type BookerInitialData,
-  type BookerStep,
-  useBooker,
-} from "@/lib/booker/use-booker";
-
-const EASE = [0.32, 0.72, 0, 1] as const;
-const SIZE_TRANSITION = { duration: 0.45, ease: EASE } as const;
-const ENTER_TRANSITION = { delay: 0.1, duration: 0.45, ease: EASE } as const;
-
-const useIsomorphicLayoutEffect =
-  typeof window === "undefined" ? useEffect : useLayoutEffect;
-
-type Size = { width: number; height: number };
-
-const measure = (element: HTMLElement): Size => ({
-  width: element.offsetWidth,
-  height: element.offsetHeight,
-});
-
-function useStepTransition(step: BookerStep) {
-  const containerRef = useRef<HTMLDivElement>(null);
-  const previousStepRef = useRef(step);
-  const restingSizeRef = useRef<Size | null>(null);
-  const isAnimatingRef = useRef(false);
-  const reducedMotion = useReducedMotion();
-
-  useIsomorphicLayoutEffect(() => {
-    const container = containerRef.current;
-    if (!container) return;
-
-    const observer = new ResizeObserver(() => {
-      if (!isAnimatingRef.current) {
-        restingSizeRef.current = measure(container);
-      }
-    });
-    observer.observe(container);
-    return () => observer.disconnect();
-  }, []);
-
-  useIsomorphicLayoutEffect(() => {
-    const container = containerRef.current;
-    if (!container || previousStepRef.current === step) return;
-    previousStepRef.current = step;
-
-    const enteringStep = container.firstElementChild as HTMLElement | null;
-
-    if (enteringStep && !reducedMotion) {
-      enteringStep.style.opacity = "0";
-    }
-
-    const start = restingSizeRef.current ?? measure(container);
-    container.style.width = "auto";
-    container.style.height = "auto";
-    enteringStep?.style.removeProperty("width");
-    void container.offsetHeight;
-    const end = measure(container);
-
-    if (enteringStep) enteringStep.style.width = `${end.width}px`;
-    container.style.width = `${start.width}px`;
-    container.style.height = `${start.height}px`;
-
-    isAnimatingRef.current = true;
-    const sizeAnimation = animate(
-      container,
-      { width: end.width, height: end.height },
-      reducedMotion ? { duration: 0 } : SIZE_TRANSITION,
-    );
-    const enterAnimation =
-      enteringStep && !reducedMotion
-        ? animate(enteringStep, { opacity: [0, 1] }, ENTER_TRANSITION)
-        : null;
-
-    void Promise.all([sizeAnimation, enterAnimation ?? Promise.resolve()]).then(
-      () => {
-        container.style.width = "";
-        container.style.height = "";
-        enteringStep?.style.removeProperty("width");
-        enteringStep?.style.removeProperty("opacity");
-        restingSizeRef.current = measure(container);
-        isAnimatingRef.current = false;
-      },
-    );
-
-    return () => {
-      sizeAnimation.stop();
-      enterAnimation?.stop();
-      isAnimatingRef.current = false;
-    };
-  }, [step, reducedMotion]);
-
-  return containerRef;
-}
-
-function BookerSteps({
-  step,
-  children,
-}: {
-  step: BookerStep;
-  children: ReactNode;
-}) {
-  const containerRef = useStepTransition(step);
-
-  return (
-    <div
-      ref={containerRef}
-      className="relative w-full @3xl:flex-1 overflow-clip"
-    >
-      {children}
-    </div>
-  );
-}
+import { type BookerInitialData, useBooker } from "@/lib/booker/use-booker";
 
 type BookerProps = {
   target: BookerTarget;
@@ -164,12 +47,6 @@ export function Booker({ initialData, target, timezone, labels }: BookerProps) {
   const headerImageAlt = displayMeta
     ? t.headerImageAlt(displayMeta.hostName, displayMeta.eventTypeTitle)
     : "Booker header";
-  const metaContentClassName = [
-    displayMeta?.eventTypeImageUrl ? "relative -mt-11 @5xl:-mt-13" : "",
-    "flex flex-col gap-6 @5xl:p-6 p-4",
-  ]
-    .filter(Boolean)
-    .join(" ");
 
   return (
     <div
@@ -183,7 +60,12 @@ export function Booker({ initialData, target, timezone, labels }: BookerProps) {
             alt={headerImageAlt}
             src={displayMeta?.eventTypeImageUrl}
           />
-          <div className={metaContentClassName}>
+          <div
+            className={cn(
+              "flex flex-col gap-6 @5xl:p-6 p-4",
+              displayMeta?.eventTypeImageUrl && "relative -mt-11 @5xl:-mt-13",
+            )}
+          >
             <div className="flex flex-col gap-4">
               <div className="flex flex-col gap-2">
                 <BookerAvatars
@@ -259,51 +141,44 @@ export function Booker({ initialData, target, timezone, labels }: BookerProps) {
             )}
           </div>
         </div>
-        <LazyMotion features={domAnimation}>
-          <MotionConfig reducedMotion="user" transition={SIZE_TRANSITION}>
-            <BookerSteps step={booker.step}>
-              {booker.step === "select" ? (
-                <div
-                  key="select"
-                  className="flex w-full @3xl:flex-row flex-col"
-                >
-                  <div className="flex @3xl:w-[min(28.75rem,100cqw-2*var(--booker-side))] w-full @3xl:shrink-0 flex-col items-center @3xl:@max-5xl:px-2 px-4 @3xl:@max-5xl:py-2 pt-3 pb-4">
-                    <BookerCalendar {...booker.calendarProps} />
-                  </div>
-                  <TimePicker
-                    labels={{
-                      hour12Short: t.hour12Short,
-                      hour24Short: t.hour24Short,
-                      noAvailableTimes: t.noAvailableTimes,
-                      noSlotsAvailable: t.noSlotsAvailable,
-                      noSlotsThisDay: t.noSlotsThisDay,
-                      noSlotsThisMonth: t.noSlotsThisMonth,
-                      use24Hour: t.use24Hour,
-                      viewFirstAvailability: t.viewFirstAvailability,
-                    }}
-                    {...booker.timePickerProps}
-                  />
-                </div>
-              ) : (
-                <div
-                  key="confirm"
-                  className="@3xl:w-[min(28.75rem,100cqw-2*var(--booker-side))] w-full @3xl:shrink-0 @3xl:@max-5xl:px-2 px-4 @3xl:@max-5xl:py-2 pt-3 pb-4"
-                >
-                  <Button variant="ghost" size="sm" onClick={booker.onBack}>
-                    <ArrowLeftIcon aria-hidden="true" />
-                    Back
-                  </Button>
-                  <Input
-                    autoFocus
-                    className="w-full"
-                    placeholder="Your name"
-                    type="text"
-                  />
-                </div>
-              )}
-            </BookerSteps>
-          </MotionConfig>
-        </LazyMotion>
+        <BookerSteps step={booker.step}>
+          {booker.step === "select" ? (
+            <div key="select" className="flex w-full @3xl:flex-row flex-col">
+              <div className="flex @3xl:w-[min(28.75rem,100cqw-2*var(--booker-side))] w-full @3xl:shrink-0 flex-col items-center @3xl:@max-5xl:px-2 px-4 @3xl:@max-5xl:py-2 pt-3 pb-4">
+                <BookerCalendar {...booker.calendarProps} />
+              </div>
+              <TimePicker
+                labels={{
+                  hour12Short: t.hour12Short,
+                  hour24Short: t.hour24Short,
+                  noAvailableTimes: t.noAvailableTimes,
+                  noSlotsAvailable: t.noSlotsAvailable,
+                  noSlotsThisDay: t.noSlotsThisDay,
+                  noSlotsThisMonth: t.noSlotsThisMonth,
+                  use24Hour: t.use24Hour,
+                  viewFirstAvailability: t.viewFirstAvailability,
+                }}
+                {...booker.timePickerProps}
+              />
+            </div>
+          ) : (
+            <div
+              key="confirm"
+              className="@3xl:w-[min(28.75rem,100cqw-2*var(--booker-side))] w-full @3xl:shrink-0 @3xl:@max-5xl:px-2 px-4 @3xl:@max-5xl:py-2 pt-3 pb-4"
+            >
+              <Button variant="ghost" size="sm" onClick={booker.onBack}>
+                <ArrowLeftIcon aria-hidden="true" />
+                Back
+              </Button>
+              <Input
+                autoFocus
+                className="w-full"
+                placeholder="Your name"
+                type="text"
+              />
+            </div>
+          )}
+        </BookerSteps>
       </Card>
       <a
         href="https://cal.com"
